+++ changelog.md
@@ -37,6 +37,7 @@ Codion Change Log
 - LoadTestPanel, add/removeSelectedItemListener() methods replaced selectedItemsObserver() method.
 ### is.codion.swing.framework.ui
 - EntityPanel, add/removeListener methods replaced with methods that return the underlying event observer.
+- EntityApplicationPanel.addInitializationListener() replaced with initializedObserver().
 ### is.codion.swing.framework.server.monitor
 - ConnectionPoolMonitoring, add/removeStatisticsListener() methods replaced statisticsObserver() method.
 - HostMonitor, add/removeServerAddedListener() and add/removeServerRemovedListener() methods replaced with serverAddedObserver() and serverRemovedObserver().